requirements.txt
pandas
Keras==2.2.5
tensorflow==2.0.0-rc0
tensorflow_probability==0.7.0
dm-sonnet==1.34
matplotlib==3.1.1
numpy==1.17.0
Pillow==6.1.0
scipy==1.3.1
SimpleITK==1.1.0
tqdm==4.23.4
seaborn==0.9.0
pytest==3.8.0
openslide-python=1.1.1
pandas=0.25.1
jupyter=1.0.0
ipython
scikit-learn
scikit-image